Move ImportUriHandler into its own file
